Add setup.py
At our organization, we love ``bash_kernel`` and want to use it on our
supercomputing node. But because of the way the repository is set up, there's
not an automated way for us to automatically build the sources as an install
file. To make matters worse, our firewall configuration doesn't let us download
the package directly from PyPI.

For that reason I've translated ``flit.ini`` into ``setup.py``. If you pull
this commit users with similar environments - including us - can follow along
the repository and deploy releases in an automated fashion without having to
access PyPI directly or copy files by hand. Version numbers, etc. are pulled
out of the package, so no further maintenance should be required.