More thoughtful completion
A variety of improvements to bash_kernel's completion, all of which try
to more semantically match what the user (read: @kdm9) expects when
completing from a given position. In particular:

- not completing command names while in something that looks like a
  path (e.g. `./path/ap<tab>` shouldn't complete the command named
  `apt`, only the path named `./path/apple.txt`).
- Splitting completion prompts at `=` and `"`, so that one can e.g. complete
  the path in `variable="./path<tab>"`.

As with all code that tries to be smart, this will likely have a few
undesirable corner cases. Please report any issues you encounter!